// frontend/models/Visibility.ts
/**
 * @license EUPL-1.2
 * SPDX-FileCopyrightText: 2022 - 2024
 * - Nils Brinckmann <nils.brinckmann@gfz-potsdam.de>
 * - Marc Hanisch <marc.hanisch@gfz-potsdam.de>
 * - Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences (GFZ, https://www.gfz-potsdam.de)
 *
 * SPDX-License-Identifier: EUPL-1.2
 */
export interface IVisible {
  // optional because configurations can't be private
  isPrivate?: boolean
  isInternal: boolean
  isPublic: boolean
}

export enum Visibility {
  Private = 'Private',
  Internal = 'Internal',
  Public = 'Public'
}